fix(zero3): fix the efficiency issue for optimizer offloading

+++ deepspeed/runtime/zero/stage3.py
@@ -1525,7 +1525,7 @@ class DeepSpeedZeroOptimizer_Stage3(ZeROOptimizer):
                     else:
                         fp32_grad_tensor = self.fp32_partitioned_groups_flat[i].grad.narrow(
                             0, dest_offset, grad_buffer.numel())
-                        fp32_grad_tensor.copy_(grad_buffer.float())
+                        fp32_grad_tensor.copy_(grad_buffer)
 
             # free the gradient
             if not get_accelerator().is_synchronized_device():
